Moved MFACheck, Captcha and parseHTML into network module
